move writer into beans factory
