+++ airbyte-integrations/connectors/destination-bigquery/src/main/kotlin/io/airbyte/integrations/destination/bigquery/BigqueryBeansFactory.kt
@@ -9,11 +9,20 @@ import com.google.cloud.bigquery.BigQuery
 import com.google.cloud.bigquery.BigQueryOptions
 import io.airbyte.cdk.load.check.DestinationCheckerSync
 import io.airbyte.cdk.load.command.DestinationCatalog
+import io.airbyte.cdk.load.orchestration.db.legacy_typing_deduping.TableCatalog
+import io.airbyte.cdk.load.orchestration.db.legacy_typing_deduping.TypingDedupingExecutionConfig
+import io.airbyte.cdk.load.orchestration.db.legacy_typing_deduping.TypingDedupingFinalTableOperations
+import io.airbyte.cdk.load.orchestration.db.legacy_typing_deduping.TypingDedupingWriter
 import io.airbyte.cdk.load.state.SyncManager
 import io.airbyte.cdk.load.task.DestinationTaskLauncher
+import io.airbyte.cdk.load.write.StreamStateStore
 import io.airbyte.cdk.load.write.WriteOperation
 import io.airbyte.integrations.destination.bigquery.check.BigqueryCheckCleaner
 import io.airbyte.integrations.destination.bigquery.spec.BigqueryConfiguration
+import io.airbyte.integrations.destination.bigquery.typing_deduping.BigQueryDatabaseHandler
+import io.airbyte.integrations.destination.bigquery.typing_deduping.BigQuerySqlGenerator
+import io.airbyte.integrations.destination.bigquery.typing_deduping.BigqueryDatabaseInitialStatusGatherer
+import io.airbyte.integrations.destination.bigquery.write.BigqueryRawTableOperations
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.micronaut.context.annotation.Factory
 import jakarta.inject.Named
@@ -39,6 +48,28 @@ class BigqueryBeansFactory {
             WriteOperation(taskLauncher, syncManager),
             BigqueryCheckCleaner(),
         )
+
+    @Singleton
+    fun getWriter(
+        bigquery: BigQuery,
+        config: BigqueryConfiguration,
+        names: TableCatalog,
+        streamStateStore: StreamStateStore<TypingDedupingExecutionConfig>,
+    ): TypingDedupingWriter {
+        val destinationHandler = BigQueryDatabaseHandler(bigquery, config.datasetLocation.region)
+        return TypingDedupingWriter(
+            names,
+            BigqueryDatabaseInitialStatusGatherer(bigquery),
+            destinationHandler,
+            BigqueryRawTableOperations(bigquery),
+            TypingDedupingFinalTableOperations(
+                BigQuerySqlGenerator(config.projectId, config.datasetLocation.region),
+                destinationHandler,
+            ),
+            disableTypeDedupe = config.disableTypingDeduping,
+            streamStateStore,
+        )
+    }
 
     @Singleton
     fun getBigqueryClient(config: BigqueryConfiguration): BigQuery {
